requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==0.7.11
bottle==0.13.3
Brotli==1.1.0
browser-cookie3==0.20.1
cairocffi==1.7.1
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
clr_loader==0.2.7.post0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
coqpit-config==0.2.0
coqui-tts==0.26.0
coqui-tts-trainer==0.2.3
cssselect2==0.8.0
curl_cffi==0.10.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
duckduckgo_search==8.0.1
EbookLib==0.19
einops==0.8.1
encodec==0.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
g4f==0.5.2.2
grpcio==1.71.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.4.0
lz4==4.4.4
magic-filter==1.0.12
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
monotonic-alignment-search==0.1.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
mss==10.0.0
multidict==6.4.3
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodriver==0.45.2
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
odfpy==1.4.1
ollama==0.4.8
openai==1.77.0
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
plyer==2.1.0
pooch==1.8.2
preshed==3.0.9
primp==0.15.0
propcache==0.3.1
protobuf==6.30.2
proxy-tools==0.1.0
psutil==7.0.0
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-socks==2.7.1
pythonnet==3.0.5
pytz==2025.2
pywebview==5.4
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
shadowcopy==0.0.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
transformers==4.51.3
typeguard==4.4.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
WMI==1.5.1
wrapt==1.17.2
yarl==1.20.0
